requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.13
beautifulsoup4>=4.12.2
requests>=2.31.0
python-dotenv>=1.0.0
google-search-results>=2.4.2
langchain-google-genai>=0.0.5
fastapi>=0.104.1
uvicorn>=0.23.2
jinja2>=3.1.2
aiosmtplib>=2.0.2
pydantic>=2.4.2
questionary>=2.0.1
colorama>=0.4.6
rich>=13.5.2
llama-cpp-python>=0.2.11
selenium>=4.10.0
webdriver-manager>=3.8.6 